subclass group chat view, some walkthrough comments
